+++ examples/seed_to_fingerprint/main.go
@@ -3,10 +3,11 @@ package main
 import (
 	"fmt"
 
-	"github.com/Gaukas/seed2sdp"
+	s2s "github.com/Gaukas/seed2sdp"
 )
 
 func main() {
-	fp, _ := seed2sdp.PredictDTLSFingerprint([]byte("ExampleSecret"), []byte("ExampleSeed"), []byte("ExamplePrefix"))
+	hkdfParams, _ := s2s.NewHKDFParams([]byte("ExampleSecret"), []byte("ExampleSeed"), []byte("ExamplePrefix"))
+	fp, _ := s2s.PredictDTLSFingerprint(hkdfParams)
 	fmt.Println("Fingerprint: ", fp)
 }
