+++ examples/seed_to_iceparams/main.go
@@ -3,11 +3,12 @@ package main
 import (
 	"fmt"
 
-	"github.com/Gaukas/seed2sdp"
+	s2s "github.com/Gaukas/seed2sdp"
 )
 
 func main() {
-	iceParams, _ := seed2sdp.PredictIceParameters([]byte("ExampleSecret"), []byte("ExampleSeed"), []byte("ExamplePrefix"))
+	hkdfParams, _ := s2s.NewHKDFParams([]byte("ExampleSecret"), []byte("ExampleSeed"), []byte("ExamplePrefix"))
+	iceParams, _ := s2s.PredictIceParameters(hkdfParams)
 	fmt.Println("ice-ufrag: ", iceParams.UsernameFragment)
 	fmt.Println("ice-pwd: ", iceParams.Password)
 	fmt.Println("ice-lite: ", iceParams.ICELite)
