+++ go.mod
@@ -8,3 +8,5 @@ require (
 	github.com/pion/webrtc/v3 v3.0.21
 	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2
 )
+
+replace github.com/Gaukas/seed2sdp => ../../../seed2sdp
